prod/pipeline/requirements.txt
azure-storage-blob>=12.0.0
azure-identity>=1.12.0
requests>=2.31.0
pandas>=2.2.3
openpyxl>=3.1.2
xlsxwriter>=3.2.2
